Refactor: move mongo connection to config main path
