+++ src/app/Application.ts
@@ -1,6 +1,6 @@
 import { CONSTANT } from '../config/constants/constants';
 import express, { Application as App } from 'express';
-import { MongoConnection } from '../db/mongoConnection';
+import { MongoConnection } from '../config/db/mongoConnection';
 import { Logger } from '../utils/logger/Logger';
 
 export class Application {
